Remove type conversion in output function

+++ baranov.pavel/T1/outputShapes.cpp
@@ -20,8 +20,8 @@ void baranov::outputShapes(std::ostream & output, Shape ** shapes, const size_t 
   {
     rectangle_t frameRect = shapes[i]->getFrameRect();
     point_t pos = frameRect.pos;
-    size_t width = frameRect.width;
-    size_t height = frameRect.height;
+    double width = frameRect.width;
+    double height = frameRect.height;
     output << pos.x - width / 2 << ' ' << pos.y - height / 2  << ' ';
     output << pos.x + width / 2 << ' ' << pos.y + height / 2;
     if (i != count - 1)
